requirements.txt
faster_whisper==1.1.1
Flask==3.1.0
gevent==24.11.1
llama_cpp_python==0.3.7